refactor : delete unused code and move common files
