+++ app/src/main/java/com/kalex/bookyouu_notesapp/common/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.kalex.bookyouu_notesapp.ui.theme
+package com.kalex.bookyouu_notesapp.common.theme
 
 import android.app.Activity
 import android.os.Build
