+++ app/src/main/java/com/kalex/bookyouu_notesapp/subject/SubjectMainScreen.kt
@@ -7,8 +7,8 @@ import com.kalex.bookyouu_notesapp.R
 import com.kalex.bookyouu_notesapp.subject.createSubject.ViewModelState
 import com.kalex.bookyouu_notesapp.subject.subjectList.presentation.SubjectListViewModel
 import com.kalex.bookyouu_notesapp.subject.subjectList.presentation.ui.SubjectListScreen
-import com.kalex.bookyouu_notesapp.ui.composables.BYLoadingIndicator
-import com.kalex.bookyouu_notesapp.ui.composables.EmptyScreen
+import com.kalex.bookyouu_notesapp.common.composables.BYLoadingIndicator
+import com.kalex.bookyouu_notesapp.common.composables.EmptyScreen
 
 @Composable
 fun SubjectMainScreen(
